+++ web/packages/new/photos/services/settings.ts
